Move index v2 semantics in seperate file
Abstraction in preparetion for index v5 implementation.
